Modularize the Triton inference model for NVT Workflows (#1252)
* Move Triton ensemble code out of init file

* Clean up the code that runs workflows during inference

This splits the Triton workflow running code into an abstract base class with the core functionality and several sub-classes that contain functionality specific to particular frameworks (e.g. Tensorflow.)

* Fix missing return and some of the tests

* Condense type conversions into dictionary lookups

* Fix some of the inference tests

* Fix imports and inits

* Add imports for generate methods

* Adjust tests for fixed imports

* Make the default model type Tensorflow
