+++ nvtabular/inference/triton/runners/hugectr.py
@@ -1,0 +1,90 @@
+# Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#  * Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+#  * Redistributions in binary form must reproduce the above copyright
+#    notice, this list of conditions and the following disclaimer in the
+#    documentation and/or other materials provided with the distribution.
+#  * Neither the name of NVIDIA CORPORATION nor the names of its
+#    contributors may be used to endorse or promote products derived
+#    from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ``AS IS'' AND ANY
+# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+# PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+# CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+# EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+# PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+# PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+# OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+import numpy as np
+from triton_python_backend_utils import InferenceResponse, Tensor
+
+from nvtabular.inference.triton import get_column_types
+from nvtabular.inference.triton.runners.base import WorkflowRunner
+from nvtabular.ops import get_embedding_sizes
+
+
+class HugeCTRWorkflowRunner(WorkflowRunner):
+    def __init__(self, workflow_path, model_kind, model_config):
+        super().__init__(workflow_path, model_kind, model_config)
+
+        self.column_types = get_column_types(self.workflow_path)
+        if "cats" in self.column_types:
+            self.offsets = self.get_offsets(self.workflow, self.column_types)
+
+    def _transform_outputs(self, tensors):
+        output_tensors = []
+        if "conts" in self.column_types:
+            output_tensors.append(
+                Tensor(
+                    "DES",
+                    self._convert(self.column_types["conts"], tensors, np.float32),
+                )
+            )
+        else:
+            output_tensors.append(Tensor("DES", np.array([[]], np.float32)))
+
+        if "cats" in self.column_types:
+            for name in self.column_types["cats"]:
+                tensors[name] += self.offsets[name]
+            cats_np = self._convert(self.column_types["cats"], tensors, np.int64)
+            output_tensors.append(
+                Tensor(
+                    "CATCOLUMN",
+                    cats_np,
+                )
+            )
+        else:
+            output_tensors.append(Tensor("CATCOLUMN", np.array([[]], np.int64)))
+
+        len_cats_np = cats_np.shape[1]
+        row_index = np.arange(len_cats_np + 1, dtype=np.int32).reshape(1, len_cats_np + 1)
+        output_tensors.append(Tensor("ROWINDEX", row_index))
+
+        return InferenceResponse(output_tensors)
+
+    def _convert(self, columns, tensors, dtype):
+        """converts outputs to a numpy input compatible with hugectr"""
+        rows = max(len(tensors[name]) for name in columns)
+        d = self._convert_to_np(columns, tensors, dtype, rows)
+        return d.reshape(1, len(columns) * rows)
+
+    def get_offsets(self, workflow, column_types):
+        embeddings = get_embedding_sizes(workflow)
+        if embeddings is None:
+            raise Exception("embeddings cannot be None")
+        else:
+            offsets = dict()
+            curr_offset = 0
+            for name in column_types["cats"]:
+                offsets[name] = curr_offset
+                curr_offset += embeddings[name][0]
+            return offsets
