+++ tests/unit/test_triton_inference.py
@@ -18,6 +18,7 @@ from tests.conftest import assert_eq
 
 triton = pytest.importorskip("nvtabular.inference.triton")
 data_conversions = pytest.importorskip("nvtabular.inference.triton.data_conversions")
+ensemble = pytest.importorskip("nvtabular.inference.triton.ensemble")
 
 grpcclient = pytest.importorskip("tritonclient.grpc")
 tritonclient = pytest.importorskip("tritonclient")
@@ -308,7 +309,7 @@ def test_remove_columns():
     df = pd.DataFrame({"a": ["a", "b"], "b": [1.0, 2.0], "label": [0, 1]})
     workflow.fit(nvt.Dataset(df))
 
-    removed = triton._remove_columns(workflow, label_columns)
+    removed = ensemble._remove_columns(workflow, label_columns)
     assert set(removed.output_dtypes.keys()) == {"a", "b"}
